backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.133.1",
    "nanoid>=2.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.47",
    "typer>=0.24.1",
    "uvicorn>=0.41.0",
]

[project.scripts]
backend = "cli.main:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[dependency-groups]
client = [
    "httpx>=0.28.1",
]
dev = [
    "pytest>=9.0.2",
]
